Move signed entity discriminant list parsing to the discriminant itself
From the aggregator configuration.
